* added logging to a lot of gm commands that needed it * removed a bunch of old/un-needed commands * moved account ban to 'a' command group
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2448 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/Object.cpp
@@ -2320,7 +2320,7 @@ void Object::SpellNonMeleeDamageLog(Unit *pVictim, uint32 spellID, uint32 damage
 	}
 	if (this->IsPlayer())
 	{
-			static_cast<Player*>(this)->m_casted_amount[school]=res;
+			static_cast<Player*>(this)->m_casted_amount[school]=(uint32)res;
 	}
 
 	DealDamage(pVictim, float2int32(res),  2, 0, spellID);
